refactor: group components to specific folders
