+++ src/App.js
@@ -4,8 +4,8 @@ import {BrowserRouter as Router,
         Switch,
         Route,
         Link} from 'react-router-dom';
-import Header from './Components/Header';
-import Footer from './Components/Footer';
+import Header from './Components/Header/Header';
+import Footer from './Components/Footer/Footer';
 import Main from './Components/Main';
 import Details from './Components/Details';
 var id;
